requirements.txt
torch
albumentations
efficientnet_pytorch
torchvision
glob
Pillow
json
tqdm